+++ src/org/usfirst/frc/team4990/robot/Robot.java
@@ -52,167 +52,7 @@ public class Robot extends TimedRobot {
 	 * Benjamin has received permission to use this code. Thank you team 199!
 	 */
 
-	public static SmartDashboardInterface sd = new SmartDashboardInterface() {
-		Preferences pref = Preferences.getInstance();
-
-		/**
-		 * Retrieves a numerical constant from SmartDashbaord/Shuffleboard.
-		 * 
-		 * @param key
-		 *            string key to identify value
-		 * @param def
-		 *            number to return if no value is retrieved
-		 * @author Deep Blue Robotics (Team 199)
-		 */
-
-		@Override
-		public double getConst(String key, double def) {
-
-			if (!pref.containsKey("Const/" + key)) {
-				pref.putDouble("Const/" + key, def);
-				if (pref.getDouble("Const/ + key", def) != def) {
-					System.err.println("pref Key" + "Const/" + key + "already taken by a different type");
-					return def;
-				}
-			}
-			return pref.getDouble("Const/" + key, def);
-		}
-
-		/**
-		 * Retrieves a string from SmartDashbaord/Shuffleboard.
-		 * 
-		 * @param key
-		 *            string key to identify value
-		 * @param def
-		 *            string to return if no value is retrieved
-		 * @author Deep Blue Robotics (Team 199)
-		 */
-
-		@Override
-		public String getString(String key, String def) {
-			if (!pref.containsKey("String/" + key)) {
-				pref.putString("String/" + key, def);
-				if (pref.getString("String/ + key", def) != def) {
-					System.err.println("pref Key" + "String/" + key + "already taken by a different type");
-					return def;
-				}
-			}
-			return pref.getString("String/" + key, def);
-		}
-
-		/**
-		 * Adds a constant to SmartDashbaord/Shuffleboard.
-		 * 
-		 * @param key
-		 *            string key to identify value
-		 * @param def
-		 *            value to be stored
-		 * @author Deep Blue Robotics (Team 199)
-		 */
-
-		@Override
-		public void putConst(String key, double def) {
-			pref.putDouble("Const/" + key, def);
-			if (pref.getDouble("Const/ + key", def) != def) {
-				System.err.println("pref Key" + "Const/" + key + "already taken by a different type");
-			}
-		}
-
-		@Override
-		public void putString(String string, String def) {
-			pref.putString("String/" + string, def);
-			if (pref.getString("String/ + key", def) != def) {
-				System.err.println("pref Key" + "String/" + def + "already taken by a different type");
-			}
-
-		}
-
-		/**
-		 * Retrieves a boolean from SmartDashbaord/Shuffleboard.
-		 * 
-		 * @param key
-		 *            string key to identify value
-		 * @param def
-		 *            boolean to return if no value is retrieved
-		 * @author Deep Blue Robotics (Team 199)
-		 */
-
-		@SuppressWarnings("unused")
-		public boolean getBool(String key, boolean def) {
-			Preferences pref = Preferences.getInstance();
-			if (!pref.containsKey("Bool/" + key)) {
-				pref.putBoolean("Bool/" + key, def);
-				if (pref.getBoolean("Bool/" + key, def) == def) {
-					System.err.println("pref Key" + "Bool/" + key + "already taken by a different type");
-					return def;
-				}
-			}
-			return pref.getBoolean("Bool/" + key, def);
-		}
-
-		@Override
-		public void putBool(String string, Boolean def) {
-			pref.putBoolean("Boolean/" + string, def);
-			if (pref.getBoolean("Boolean/ + key", def) != def) {
-				System.err.println("pref Key" + "Boolean/" + def + "already taken by a different type");
-			}
-
-		}
-	};
-
-	/**
-	 * Retrieves a numerical constant from SmartDashbaord/Shuffleboard.
-	 * 
-	 * @param key
-	 *            string key to identify value
-	 * @param def
-	 *            number to return if no value is retrieved
-	 * @author Deep Blue Robotics (Team 199)
-	 */
-
-	public static double getConst(String key, double def) {
-		return sd.getConst(key, def);
-	}
-
-	/**
-	 * Retrieves a string from SmartDashbaord/Shuffleboard.
-	 * 
-	 * @param key
-	 *            string key to identify value
-	 * @param def
-	 *            string to return if no value is retrieved
-	 * @author Deep Blue Robotics (Team 199)
-	 */
-
-	public static String getString(String key, String def) {
-		return sd.getString(key, def);
-	}
-
-	public static void putConst(String key, double def) {
-		sd.putConst(key, def);
-	}
-
-	/**
-	 * Retrieves a boolean from SmartDashbaord/Shuffleboard.
-	 * 
-	 * @param key
-	 *            string key to identify value
-	 * @param def
-	 *            boolean to return if no value is retrieved
-	 * @author Deep Blue Robotics (Team 199)
-	 */
-
-	public static boolean getBool(String key, boolean def) {
-		Preferences pref = Preferences.getInstance();
-		if (!pref.containsKey("Bool/" + key)) {
-			pref.putBoolean("Bool/" + key, def);
-			if (pref.getBoolean("Bool/" + key, def) == def) {
-				System.err.println("pref Key" + "Bool/" + key + "already taken by a different type");
-				return def;
-			}
-		}
-		return pref.getBoolean("Bool/" + key, def);
-	}
+	public static SmartDashboardInterface sd = new SmartDashboardInterface();
 
 	/**
 	 * @author Team 4990
@@ -420,5 +260,5 @@ public class Robot extends TimedRobot {
 	// ever heard of the tale of last minute code
 	// I thought not, it is not a tale the chairman will tell to you
 	// (Keep below last function of Robot.java)
-
+	
 }
